[CHANGE] Moved card commands to a separate file
